Tidy up license headers
Remove white spaces at the end of the lines and in empty lines and
remove last empty line (not part of the header).
Replace the year with a token, so that it does not require update the
headers each year (it's up to the tool or user to replace it with the
current one). Update README to mention that.