Update Gradle and migrate to Kotlin DSL
Update Gradle to 4.8 and Spotless plugin to 3.13.0.
Change the build scripts to Kotlin DSL and extract a task to reduce some
code duplication when declaring the ZAP tasks (e.g. group, dependencies,
and classpath).
Copy the ZapVersions to a build directory instead of including them from
the root dir, for proper inclusion of default test resources along with
ZapVersions files.