Download released add-ons
Add task and update build to download the add-ons released in the latest
commit, to be able to generate the website pages.
Add list of allowed add-ons to only download the ones required (e.g.
exclude 3rd-party add-ons without help).
Extract the download of the add-on and checksum calculation to a utility
class.

Signed-off-by: thc202 <thc202@gmail.com>