Update generate_alert_pages.js script
`getExampleAlerts` is a public method for most scan rules and reflection
is not needed to access it.

The reason this change was needed is that reflection was failing for
`PassiveScriptScanRule`s. This is probably because we are using
`ByteBuddy` to dynamically create a unique class for each passive script
but I haven't checked in detail.

Signed-off-by: ricekot <github@ricekot.com>